+++ CMS/CMS/DataTracking/DataIO.cs
@@ -9,7 +9,7 @@ using System.Data.SqlClient;
 using DataControlsLib;
 using DataControlsLib.DataModels;
 
-namespace CMS
+namespace CMS.DataTracking
 {
     public class DataIO
     {
@@ -22,17 +22,17 @@ namespace CMS
             {
                 GetDB.GetDataTable(connection, ds_io, "tblDataIORequests",
                     "SELECT RequestID, Project, ChangeDate, ChangeType, RequestedBy, ChangedBy FROM dbo.tblDataIORequests ORDER BY ChangeDate DESC");
-                GetDB.GetDataTable(connection, ds_io, "tblAssetsRegister", 
+                GetDB.GetDataTable(connection, ds_io, "tblAssetsRegister",
                     "SELECT AssetID, AssetName, AssetSha256sum, VreFilePath FROM dbo.tblAssetsRegister");
                 GetDB.GetDataTable(connection, ds_io, "tblAssetsChangeLog",
                     "SELECT ChangeID, RequestID, AssetID, ChangeAccepted, RejectionNotes FROM dbo.tblAssetsChangeLog");
-                GetDB.GetDataTable(connection, ds_io, "tlkAssetChangeTypes", 
+                GetDB.GetDataTable(connection, ds_io, "tlkAssetChangeTypes",
                     "SELECT ChangeTypeID, ChangeTypeLabel FROM dbo.tlkAssetChangeTypes");
                 GetDB.GetDataTable(connection, ds_io, "tblProject",
                     "SELECT * FROM dbo.tblProject WHERE ValidTo IS NULL");
 
                 // Relations not needed due to LINQ joins, but at least they help enforce the schema
-                ds_io.Relations.Add("DataIORequests_AssetChangeTypes", 
+                ds_io.Relations.Add("DataIORequests_AssetChangeTypes",
                     parentColumn: ds_io.Tables["tlkAssetChangeTypes"].Columns["ChangeTypeID"],
                     childColumn: ds_io.Tables["tblDataIORequests"].Columns["ChangeType"]);
                 ds_io.Relations.Add("AssetsChangeLog_DataIORequests",
